56 Holmefield Avenue
"Indeed! My mistress told me that you were likely to call. She left this morning with her husband by
the 5.15 train from Charing Cross for the Continent."

"What!" Sherlock Holmes staggered back, white with chagrin and surprise. "Do you mean that she has
left England?"

"Never to return."

"And the papers?" asked the King hoarsely. "All is lost."

"We shall see." He pushed past the servant and rushed into the drawing-room, followed by the King
and myself. The furniture was scattered about in every direction, with dismantled shelves and open
drawers, as if the lady had hurriedly ransacked them before her flight. Holmes rushed at the bell-
pull, tore back a small sliding shutter, and, plunging in his hand, pulled out a photograph and a
letter. The photograph was of Irene Adler herself in evening dress, the letter was superscribed to
"Sherlock Holmes, Esq. To be left till called for." My friend tore it open and we all three read it
together. It was dated at midnight of the preceding night and ran in this way: