9 Tamworth Drive
"It was only a joke at first," groaned our visitor. "We never thought that she would have been so
carried away."

"Very likely not. However that may be, the young lady was very decidedly carried away, and, having
quite made up her mind that her stepfather was in France, the suspicion of treachery never for an
instant entered her mind. She was flattered by the gentleman's attentions, and the effect was
increased by the loudly expressed admiration of her mother. Then Mr. Angel began to call, for it was
obvious that the matter should be pushed as far as it would go if a real effect were to be produced.
There were meetings, and an engagement, which would finally secure the girl's affections from
turning towards anyone else. But the deception could not be kept up forever. These pretended
journeys to France were rather cumbrous. The thing to do was clearly to bring the business to an end
in such a dramatic manner that it would leave a permanent impression upon the young lady's mind and
prevent her from looking upon any other suitor for some time to come. Hence those vows of fidelity
exacted upon a Testament, and hence also the allusions to a possibility of something happening on
the very morning of the wedding. James Windibank wished Miss Sutherland to be so bound to Hosmer
Angel, and so uncertain as to his fate, that for ten years to come, at any rate, she would not
listen to another man. As far as the church door he brought her, and then, as he could go no
farther, he conveniently vanished away by the old trick of stepping in at one door of a four-wheeler
and out at the other. I think that was the chain of events, Mr. Windibank!"