53 Down Close
"As it happens, he spoke of coming into town to-day upon some most important business. It is
probable that he will be away all day, and that there would be nothing to disturb you. We have a
housekeeper now, but she is old and foolish, and I could easily get her out of the way."

"Excellent. You are not averse to this trip, Watson?"