4 Tamworth Drive
"They are typewritten," I remarked.

"Not only that, but the signature is typewritten. Look at the neat little 'Hosmer Angel' at the
bottom. There is a date, you see, but no superscription except Leadenhall Street, which is rather
vague. The point about the signature is very suggestive--in fact, we may call it conclusive."

"Of what?"

"My dear fellow, is it possible you do not see how strongly it bears upon the case?"

"I cannot say that I do unless it were that he wished to be able to deny his signature if an action
for breach of promise were instituted."