133 Badgers Dene
"And the ring?" I asked, glancing at a remarkable brilliant which sparkled upon his finger.

"It was from the reigning family of Holland, though the matter in which I served them was of such
delicacy that I cannot confide it even to you, who have been good enough to chronicle one or two of
my little problems."

"And have you any on hand just now?" I asked with interest.

"Some ten or twelve, but none which present any feature of interest. They are important, you
understand, without being interesting. Indeed, I have found that it is usually in unimportant
matters that there is a field for the observation, and for the quick analysis of cause and effect
which gives the charm to an investigation. The larger crimes are apt to be the simpler, for the
bigger the crime the more obvious, as a rule, is the motive. In these cases, save for one rather
intricate matter which has been referred to me from Marseilles, there is nothing which presents any
features of interest. It is possible, however, that I may have something better before very many
minutes are over, for this is one of my clients, or I am much mistaken."