147 Larch Walk
"Which were very well justified," observed Holmes. "And now it is time that we arranged our little
plans. I expect that within an hour matters will come to a head. In the meantime Mr. Merryweather,
we must put the screen over that dark lantern."

"And sit in the dark?"

"I am afraid so. I had brought a pack of cards in my pocket, and I thought that, as we were a partie
carrée, you might have your rubber after all. But I see that the enemy's preparations have gone so
far that we cannot risk the presence of a light. And, first of all, we must choose our positions.
These are daring men, and though we shall take them at a disadvantage, they may do us some harm
unless we are careful. I shall stand behind this crate, and do you conceal yourselves behind those.
Then, when I flash a light upon them, close in swiftly. If they fire, Watson, have no compunction
about shooting them down."

I placed my revolver, cocked, upon the top of the wooden case behind which I crouched. Holmes shot
the slide across the front of his lantern and left us in pitch darkness--such an absolute darkness
as I have never before experienced. The smell of hot metal remained to assure us that the light was
still there, ready to flash out at a moment's notice. To me, with my nerves worked up to a pitch of
expectancy, there was something depressing and subduing in the sudden gloom, and in the cold dank
air of the vault.

"They have but one retreat," whispered Holmes. "That is back through the house into Saxe-Coburg
Square. I hope that you have done what I asked you, Jones?"