17 Badgers Dene
"Yes, and I have no doubt that is her ring."

"It may turn out to be of more interest than you think. You remember that the affair of the blue
carbuncle, which appeared to be a mere whim at first, developed into a serious investigation. It may
be so in this case, also."

"Well, let us hope so. But our doubts will very soon be solved, for here, unless I am much mistaken,
is the person in question."

As he spoke the door opened and a young lady entered the room. She was plainly but neatly dressed,
with a bright, quick face, freckled like a plover's egg, and with the brisk manner of a woman who
has had her own way to make in the world.

"You will excuse my troubling you, I am sure," said she, as my companion rose to greet her, "but I
have had a very strange experience, and as I have no parents or relations of any sort from whom I
could ask advice, I thought that perhaps you would be kind enough to tell me what I should do."