151 Larch Walk
"Well, it was the beginning of '84 when my father came to live at Horsham, and all went as well as
possible with us until the January of '85. On the fourth day after the new year I heard my father
give a sharp cry of surprise as we sat together at the breakfast-table. There he was, sitting with a
newly opened envelope in one hand and five dried orange pips in the outstretched palm of the other
one. He had always laughed at what he called my cock-and-bull story about the colonel, but he looked
very scared and puzzled now that the same thing had come upon himself.

"'Why, what on earth does this mean, John?' he stammered.