37 Valentia Road
"And one who was at home in the district, for the pool can only be approached by the farm or by the
estate, where strangers could hardly wander."

"Quite so."

"Then comes our expedition of to-day. By an examination of the ground I gained the trifling details
which I gave to that imbecile Lestrade, as to the personality of the criminal."

"But how did you gain them?"

"You know my method. It is founded upon the observation of trifles."