73 Bowler Street
"Well," said our engineer ruefully as we took our seats to return once more to London, "it has been
a pretty business for me! I have lost my thumb and I have lost a fifty-guinea fee, and what have I
gained?"

"Experience," said Holmes, laughing. "Indirectly it may be of value, you know; you have only to put
it into words to gain the reputation of being excellent company for the remainder of your
existence."

The Lord St. Simon marriage, and its curious termination, have long ceased to be a subject of
interest in those exalted circles in which the unfortunate bridegroom moves. Fresh scandals have
eclipsed it, and their more piquant details have drawn the gossips away from this four-year-old
drama. As I have reason to believe, however, that the full facts have never been revealed to the
general public, and as my friend Sherlock Holmes had a considerable share in clearing the matter up,
I feel that no memoir of him would be complete without some little sketch of this remarkable
episode.

It was a few weeks before my own marriage, during the days when I was still sharing rooms with
Holmes in Baker Street, that he came home from an afternoon stroll to find a letter on the table
waiting for him. I had remained indoors all day, for the weather had taken a sudden turn to rain,
with high autumnal winds, and the Jezail bullet which I had brought back in one of my limbs as a
relic of my Afghan campaign throbbed with dull persistence. With my body in one easy-chair and my
legs upon another, I had surrounded myself with a cloud of newspapers until at last, saturated with
the news of the day, I tossed them all aside and lay listless, watching the huge crest and monogram
upon the envelope upon the table and wondering lazily who my friend's noble correspondent could be.