86 Badgers Dene
The ejaculation had been drawn from my companion by the fact that our door had been suddenly dashed
open, and that a huge man had framed himself in the aperture. His costume was a peculiar mixture of
the professional and of the agricultural, having a black top-hat, a long frock-coat, and a pair of
high gaiters, with a hunting-crop swinging in his hand. So tall was he that his hat actually brushed
the cross bar of the doorway, and his breadth seemed to span it across from side to side. A large
face, seared with a thousand wrinkles, burned yellow with the sun, and marked with every evil
passion, was turned from one to the other of us, while his deep-set, bile-shot eyes, and his high,
thin, fleshless nose, gave him somewhat the resemblance to a fierce old bird of prey.

"Which of you is Holmes?" asked this apparition.

"My name, sir; but you have the advantage of me," said my companion quietly.

"I am Dr. Grimesby Roylott, of Stoke Moran."

"Indeed, Doctor," said Holmes blandly. "Pray take a seat."