103 Beaconside
"Away they went, and I was just wondering whether I should not do well to follow them when up the
lane came a neat little landau, the coachman with his coat only half-buttoned, and his tie under his
ear, while all the tags of his harness were sticking out of the buckles. It hadn't pulled up before
she shot out of the hall door and into it. I only caught a glimpse of her at the moment, but she was
a lovely woman, with a face that a man might die for.

"'The Church of St. Monica, John,' she cried, 'and half a sovereign if you reach it in twenty
minutes.'

"This was quite too good to lose, Watson. I was just balancing whether I should run for it, or
whether I should perch behind her landau when a cab came through the street. The driver looked twice
at such a shabby fare, but I jumped in before he could object. 'The Church of St. Monica,' said I,
'and half a sovereign if you reach it in twenty minutes.' It was twenty-five minutes to twelve, and
of course it was clear enough what was in the wind.