66 Glan Road
"The next I heard of Frank was that he was in Montana, and then he went prospecting in Arizona, and
then I heard of him from New Mexico. After that came a long newspaper story about how a miners' camp
had been attacked by Apache Indians, and there was my Frank's name among the killed. I fainted dead
away, and I was very sick for months after. Pa thought I had a decline and took me to half the
doctors in 'Frisco. Not a word of news came for a year and more, so that I never doubted that Frank
was really dead. Then Lord St. Simon came to 'Frisco, and we came to London, and a marriage was
arranged, and pa was very pleased, but I felt all the time that no man on this earth would ever take
the place in my heart that had been given to my poor Frank.

"Still, if I had married Lord St. Simon, of course I'd have done my duty by him. We can't command
our love, but we can our actions. I went to the altar with him with the intention to make him just
as good a wife as it was in me to be. But you may imagine what I felt when, just as I came to the
altar rails, I glanced back and saw Frank standing and looking at me out of the first pew. I thought
it was his ghost at first; but when I looked again there he was still, with a kind of question in
his eyes, as if to ask me whether I were glad or sorry to see him. I wonder I didn't drop. I know
that everything was turning round, and the words of the clergyman were just like the buzz of a bee
in my ear. I didn't know what to do. Should I stop the service and make a scene in the church? I
glanced at him again, and he seemed to know what I was thinking, for he raised his finger to his
lips to tell me to be still. Then I saw him scribble on a piece of paper, and I knew that he was
writing me a note. As I passed his pew on the way out I dropped my bouquet over to him, and he
slipped the note into my hand when he returned me the flowers. It was only a line asking me to join
him when he made the sign to me to do so. Of course I never doubted for a moment that my first duty
was now to him, and I determined to do just whatever he might direct.

"When I got back I told my maid, who had known him in California, and had always been his friend. I
ordered her to say nothing, but to get a few things packed and my ulster ready. I know I ought to
have spoken to Lord St. Simon, but it was dreadful hard before his mother and all those great
people. I just made up my mind to run away and explain afterwards. I hadn't been at the table ten
minutes before I saw Frank out of the window at the other side of the road. He beckoned to me and
then began walking into the Park. I slipped out, put on my things, and followed him. Some woman came
talking something or other about Lord St. Simon to me--seemed to me from the little I heard as if he
had a little secret of his own before marriage also--but I managed to get away from her and soon
overtook Frank. We got into a cab together, and away we drove to some lodgings he had taken in
Gordon Square, and that was my true wedding after all those years of waiting. Frank had been a
prisoner among the Apaches, had escaped, came on to 'Frisco, found that I had given him up for dead
and had gone to England, followed me there, and had come upon me at last on the very morning of my
second wedding."

"I saw it in a paper," explained the American. "It gave the name and the church but not where the
lady lived."