107 Holmefield Avenue
He snatched it from her in his eagerness, and smoothing it out upon the table he drew over the lamp
and examined it intently. I had left my chair and was gazing at it over his shoulder. The envelope
was a very coarse one and was stamped with the Gravesend postmark and with the date of that very
day, or rather of the day before, for it was considerably after midnight.

"Coarse writing," murmured Holmes. "Surely this is not your husband's writing, madam."

"No, but the enclosure is."

"I perceive also that whoever addressed the envelope had to go and inquire as to the address."

"How can you tell that?"

"The name, you see, is in perfectly black ink, which has dried itself. The rest is of the greyish
colour, which shows that blotting-paper has been used. If it had been written straight off, and then
blotted, none would be of a deep black shade. This man has written the name, and there has then been
a pause before he wrote the address, which can only mean that he was not familiar with it. It is, of
course, a trifle, but there is nothing so important as trifles. Let us now see the letter. Ha!
there has been an enclosure here!"