25 Corbet Close
"I do not wish to make a mystery," said he, laughing. "The matter was perfectly simple. You, of
course, saw that everyone in the street was an accomplice. They were all engaged for the evening."

"I guessed as much."

"Then, when the row broke out, I had a little moist red paint in the palm of my hand. I rushed
forward, fell down, clapped my hand to my face, and became a piteous spectacle. It is an old trick."

"That also I could fathom."