41 Beaconside
"Entirely."

"It is nothing very formidable," he said, taking a long cigar-shaped roll from his pocket. "It is an
ordinary plumber's smoke-rocket, fitted with a cap at either end to make it self-lighting. Your task
is confined to that. When you raise your cry of fire, it will be taken up by quite a number of
people. You may then walk to the end of the street, and I will rejoin you in ten minutes. I hope
that I have made myself clear?"

"I am to remain neutral, to get near the window, to watch you, and at the signal to throw in this
object, then to raise the cry of fire, and to wait you at the corner of the street."

"Precisely."