11 Ventnor Terrace
"My dear fellow," said Sherlock Holmes as we sat on either side of the fire in his lodgings at Baker
Street, "life is infinitely stranger than anything which the mind of man could invent. We would not
dare to conceive the things which are really mere commonplaces of existence. If we could fly out of
that window hand in hand, hover over this great city, gently remove the roofs, and peep in at the
queer things which are going on, the strange coincidences, the plannings, the cross-purposes, the
wonderful chains of events, working through generations, and leading to the most outré results, it
would make all fiction with its conventionalities and foreseen conclusions most stale and
unprofitable."

"And yet I am not convinced of it," I answered. "The cases which come to light in the papers are, as
a rule, bald enough, and vulgar enough. We have in our police reports realism pushed to its extreme
limits, and yet the result is, it must be confessed, neither fascinating nor artistic."

"A certain selection and discretion must be used in producing a realistic effect," remarked Holmes.
"This is wanting in the police report, where more stress is laid, perhaps, upon the platitudes of
the magistrate than upon the details, which to an observer contain the vital essence of the whole
matter. Depend upon it, there is nothing so unnatural as the commonplace."

I smiled and shook my head. "I can quite understand your thinking so." I said. "Of course, in your
position of unofficial adviser and helper to everybody who is absolutely puzzled, throughout three
continents, you are brought in contact with all that is strange and bizarre. But here"--I picked up
the morning paper from the ground--"let us put it to a practical test. Here is the first heading
upon which I come. 'A husband's cruelty to his wife.' There is half a column of print, but I know
without reading it that it is all perfectly familiar to me. There is, of course, the other woman,
the drink, the push, the blow, the bruise, the sympathetic sister or landlady. The crudest of
writers could invent nothing more crude."

"Indeed, your example is an unfortunate one for your argument," said Holmes, taking the paper and
glancing his eye down it. "This is the Dundas separation case, and, as it happens, I was engaged in
clearing up some small points in connection with it. The husband was a teetotaler, there was no
other woman, and the conduct complained of was that he had drifted into the habit of winding up
every meal by taking out his false teeth and hurling them at his wife, which, you will allow, is not
an action likely to occur to the imagination of the average story-teller. Take a pinch of snuff,
Doctor, and acknowledge that I have scored over you in your example."

He held out his snuffbox of old gold, with a great amethyst in the centre of the lid. Its splendour
was in such contrast to his homely ways and simple life that I could not help commenting upon it.