76 Moor End
"It was pitch dark inside the house, and the colonel fumbled about looking for matches and muttering
under his breath. Suddenly a door opened at the other end of the passage, and a long, golden bar of
light shot out in our direction. It grew broader, and a woman appeared with a lamp in her hand,
which she held above her head, pushing her face forward and peering at us. I could see that she was
pretty, and from the gloss with which the light shone upon her dark dress I knew that it was a rich
material. She spoke a few words in a foreign tongue in a tone as though asking a question, and when
my companion answered in a gruff monosyllable she gave such a start that the lamp nearly fell from
her hand. Colonel Stark went up to her, whispered something in her ear, and then, pushing her back
into the room from whence she had come, he walked towards me again with the lamp in his hand.

"'Perhaps you will have the kindness to wait in this room for a few minutes,' said he, throwing open
another door. It was a quiet, little, plainly furnished room, with a round table in the centre, on
which several German books were scattered. Colonel Stark laid down the lamp on the top of a
harmonium beside the door. 'I shall not keep you waiting an instant,' said he, and vanished into the
darkness.

"I glanced at the books upon the table, and in spite of my ignorance of German I could see that two
of them were treatises on science, the others being volumes of poetry. Then I walked across to the
window, hoping that I might catch some glimpse of the country-side, but an oak shutter, heavily
barred, was folded across it. It was a wonderfully silent house. There was an old clock ticking
loudly somewhere in the passage, but otherwise everything was deadly still. A vague feeling of
uneasiness began to steal over me. Who were these German people, and what were they doing living in
this strange, out-of-the-way place? And where was the place? I was ten miles or so from Eyford, that
was all I knew, but whether north, south, east, or west I had no idea. For that matter, Reading, and
possibly other large towns, were within that radius, so the place might not be so secluded, after
all. Yet it was quite certain, from the absolute stillness, that we were in the country. I paced up
and down the room, humming a tune under my breath to keep up my spirits and feeling that I was
thoroughly earning my fifty-guinea fee.

"Suddenly, without any preliminary sound in the midst of the utter stillness, the door of my room
swung slowly open. The woman was standing in the aperture, the darkness of the hall behind her, the
yellow light from my lamp beating upon her eager and beautiful face. I could see at a glance that
she was sick with fear, and the sight sent a chill to my own heart. She held up one shaking finger
to warn me to be silent, and she shot a few whispered words of broken English at me, her eyes
glancing back, like those of a frightened horse, into the gloom behind her.

"'I would go,' said she, trying hard, as it seemed to me, to speak calmly; 'I would go. I should not
stay here. There is no good for you to do.'

"'But, madam,' said I, 'I have not yet done what I came for. I cannot possibly leave until I have
seen the machine.'