94 Beaconside
"You are right," said Holmes demurely; "you do find it very hard to tackle the facts."

"Anyhow, I have grasped one fact which you seem to find it difficult to get hold of," replied
Lestrade with some warmth.

"And that is--"

"That McCarthy senior met his death from McCarthy junior and that all theories to the contrary are
the merest moonshine."

"Well, moonshine is a brighter thing than fog," said Holmes, laughing. "But I am very much mistaken
if this is not Hatherley Farm upon the left."

"Yes, that is it." It was a widespread, comfortable-looking building, two-storied, slate-roofed,
with great yellow blotches of lichen upon the grey walls. The drawn blinds and the smokeless
chimneys, however, gave it a stricken look, as though the weight of this horror still lay heavy upon
it. We called at the door, when the maid, at Holmes' request, showed us the boots which her master
wore at the time of his death, and also a pair of the son's, though not the pair which he had then
had. Having measured these very carefully from seven or eight different points, Holmes desired to be
led to the court-yard, from which we all followed the winding track which led to Boscombe Pool.