72 Valentia Road
"No good news?"

"None."

"No bad?"

"No."

"Thank God for that. But come in. You must be weary, for you have had a long day."

"This is my friend, Dr. Watson. He has been of most vital use to me in several of my cases, and a
lucky chance has made it possible for me to bring him out and associate him with this
investigation."