93 Larch Walk
"We must sit without light. He would see it through the ventilator."

I nodded again.

"Do not go asleep; your very life may depend upon it. Have your pistol ready in case we should need
it. I will sit on the side of the bed, and you in that chair."

I took out my revolver and laid it on the corner of the table.

Holmes had brought up a long thin cane, and this he placed upon the bed beside him. By it he laid
the box of matches and the stump of a candle. Then he turned down the lamp, and we were left in
darkness.