98 Larch Walk
"That is the letter which I have just received, Mr. Holmes, and my mind is made up that I will
accept it. I thought, however, that before taking the final step I should like to submit the whole
matter to your consideration."

"Well, Miss Hunter, if your mind is made up, that settles the question," said Holmes, smiling.

"But you would not advise me to refuse?"