48 Down Close
"How is that?" I asked when I had finished.

"Capital! Between your brandy and your bandage, I feel a new man. I was very weak, but I have had a
good deal to go through."

"Perhaps you had better not speak of the matter. It is evidently trying to your nerves."

"Oh, no, not now. I shall have to tell my tale to the police; but, between ourselves, if it were not
for the convincing evidence of this wound of mine, I should be surprised if they believed my
statement, for it is a very extraordinary one, and I have not much in the way of proof with which to
back it up; and, even if they believe me, the clues which I can give them are so vague that it is a
question whether justice will be done."