4 Bowler Street
"What else can be indicated by that right cuff so very shiny for five inches, and the left one with
the smooth patch near the elbow where you rest it upon the desk?"

"Well, but China?"

"The fish that you have tattooed immediately above your right wrist could only have been done in
China. I have made a small study of tattoo marks and have even contributed to the literature of the
subject. That trick of staining the fishes' scales of a delicate pink is quite peculiar to China.
When, in addition, I see a Chinese coin hanging from your watch-chain, the matter becomes even more
simple."