44 Tamworth Drive
"It surprised me."

"But, surely, it was obvious. I was then much surprised and interested on glancing down to observe
that, though the boots which she was wearing were not unlike each other, they were really odd ones;
the one having a slightly decorated toe-cap, and the other a plain one. One was buttoned only in the
two lower buttons out of five, and the other at the first, third, and fifth. Now, when you see that
a young lady, otherwise neatly dressed, has come away from home with odd boots, half-buttoned, it is
no great deduction to say that she came away in a hurry."