118 Badgers Dene
"Do you know him?"

"Oh, yes! he is the green-grocer who brings our vegetables round.  His name is Francis Prosper."

"He stood," said Holmes, "to the left of the door--that is to say, farther up the path than is
necessary to reach the door?"

"Yes, he did."

"And he is a man with a wooden leg?"