55 Wellswood Gardens
"I had," said he, "come to an entirely erroneous conclusion which shows, my dear Watson, how
dangerous it always is to reason from insufficient data. The presence of the gipsies, and the use of
the word 'band,' which was used by the poor girl, no doubt, to explain the appearance which she had
caught a hurried glimpse of by the light of her match, were sufficient to put me upon an entirely
wrong scent. I can only claim the merit that I instantly reconsidered my position when, however, it
became clear to me that whatever danger threatened an occupant of the room could not come either
from the window or the door. My attention was speedily drawn, as I have already remarked to you, to
this ventilator, and to the bell-rope which hung down to the bed. The discovery that this was a
dummy, and that the bed was clamped to the floor, instantly gave rise to the suspicion that the rope
was there as a bridge for something passing through the hole and coming to the bed. The idea of a
snake instantly occurred to me, and when I coupled it with my knowledge that the doctor was
furnished with a supply of creatures from India, I felt that I was probably on the right track. The
idea of using a form of poison which could not possibly be discovered by any chemical test was just
such a one as would occur to a clever and ruthless man who had had an Eastern training. The rapidity
with which such a poison would take effect would also, from his point of view, be an advantage. It
would be a sharp-eyed coroner, indeed, who could distinguish the two little dark punctures which
would show where the poison fangs had done their work. Then I thought of the whistle. Of course he
must recall the snake before the morning light revealed it to the victim. He had trained it,
probably by the use of the milk which we saw, to return to him when summoned. He would put it
through this ventilator at the hour that he thought best, with the certainty that it would crawl
down the rope and land on the bed. It might or might not bite the occupant, perhaps she might escape
every night for a week, but sooner or later she must fall a victim.

"I had come to these conclusions before ever I had entered his room. An inspection of his chair
showed me that he had been in the habit of standing on it, which of course would be necessary in
order that he should reach the ventilator. The sight of the safe, the saucer of milk, and the loop
of whipcord were enough to finally dispel any doubts which may have remained. The metallic clang
heard by Miss Stoner was obviously caused by her stepfather hastily closing the door of his safe
upon its terrible occupant. Having once made up my mind, you know the steps which I took in order to
put the matter to the proof. I heard the creature hiss as I have no doubt that you did also, and I
instantly lit the light and attacked it."

"With the result of driving it through the ventilator."

"And also with the result of causing it to turn upon its master at the other side. Some of the blows
of my cane came home and roused its snakish temper, so that it flew upon the first person it saw. In
this way I am no doubt indirectly responsible for Dr. Grimesby Roylott's death, and I cannot say
that it is likely to weigh very heavily upon my conscience."