92 Tamworth Drive
"My dear fellow, is it possible you do not see how strongly it bears upon the case?"

"I cannot say that I do unless it were that he wished to be able to deny his signature if an action
for breach of promise were instituted."

"No, that was not the point. However, I shall write two letters, which should settle the matter. One
is to a firm in the City, the other is to the young lady's stepfather, Mr. Windibank, asking him
whether he could meet us here at six o'clock tomorrow evening. It is just as well that we should do
business with the male relatives. And now, Doctor, we can do nothing until the answers to those
letters come, so we may put our little problem upon the shelf for the interim."

I had had so many reasons to believe in my friend's subtle powers of reasoning and extraordinary
energy in action that I felt that he must have some solid grounds for the assured and easy demeanour
with which he treated the singular mystery which he had been called upon to fathom. Once only had I
known him to fail, in the case of the King of Bohemia and of the Irene Adler photograph; but when I
looked back to the weird business of the Sign of Four, and the extraordinary circumstances connected
with the Study in Scarlet, I felt that it would be a strange tangle indeed which he could not
unravel.

I left him then, still puffing at his black clay pipe, with the conviction that when I came again on
the next evening I would find that he held in his hands all the clues which would lead up to the
identity of the disappearing bridegroom of Miss Mary Sutherland.

A professional case of great gravity was engaging my own attention at the time, and the whole of
next day I was busy at the bedside of the sufferer. It was not until close upon six o'clock that I
found myself free and was able to spring into a hansom and drive to Baker Street, half afraid that I
might be too late to assist at the dénouement of the little mystery. I found Sherlock Holmes alone,
however, half asleep, with his long, thin form curled up in the recesses of his armchair. A
formidable array of bottles and test-tubes, with the pungent cleanly smell of hydrochloric acid,
told me that he had spent his day in the chemical work which was so dear to him.