113 Beaconside
"'Mr. Hatherley?' said he, with something of a German accent. 'You have been recommended to me, Mr.
Hatherley, as being a man who is not only proficient in his profession but is also discreet and
capable of preserving a secret.'

"I bowed, feeling as flattered as any young man would at such an address. 'May I ask who it was who
gave me so good a character?'

"'Well, perhaps it is better that I should not tell you that just at this moment. I have it from the
same source that you are both an orphan and a bachelor and are residing alone in London.'

"'That is quite correct,' I answered; 'but you will excuse me if I say that I cannot see how all
this bears upon my professional qualifications. I understand that it was on a professional matter
that you wished to speak to me?'

"'Undoubtedly so. But you will find that all I say is really to the point. I have a professional
commission for you, but absolute secrecy is quite essential--absolute secrecy, you understand, and
of course we may expect that more from a man who is alone than from one who lives in the bosom of
his family.'