61 Beaconside
"And the cigar-holder?"

"I could see that the end had not been in his mouth. Therefore he used a holder. The tip had been
cut off, not bitten off, but the cut was not a clean one, so I deduced a blunt pen-knife."

"Holmes," I said, "you have drawn a net round this man from which he cannot escape, and you have
saved an innocent human life as truly as if you had cut the cord which was hanging him. I see the
direction in which all this points. The culprit is--"

"Mr. John Turner," cried the hotel waiter, opening the door of our sitting-room, and ushering in a
visitor.